requirements.txt
flask
requests
dash
python-dotenv
